// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  http: typeof http;
  rls: typeof rls;
  tasks: typeof tasks;
}>;
declare const fullApiWithMounts: typeof fullApi;

export declare const api: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "public">
>;